fix (CheckmateDetector.java) corrected error for int to byte lossy conversion while assiging 6 or -6 or 1 or 0 to byte varibale

+++ TMChessServer/src/main/java/com/thinking/machines/chess/server/logic/CheckmateDetector.java
@@ -82,7 +82,7 @@ dummyBoard[e][f]=dummyTile;
 int row;
 int column;
 ArrayList<PieceMoves> capturingPiecesMovesList;
-byte pieceColor=(sourcePiece>0)?1:0;
+byte pieceColor=(byte)((sourcePiece>0)?1:0);
 for(PossibleMovesIndex pmi:piecesValidIndexes)
 {
 row=pmi.row;
